test(pns): the ported suite splits by concern and signs independently
The dispatch assertions parse events as JSON rather than substrings,
the native phases live in their own file over a shared sandbox, and
the hermes signature check computes its own HMAC with the crates
directly instead of calling the engine's signer, because an
expectation produced by the code under test is a tautology: a broken
signer would agree with itself. Mutation-proved, one byte into the
signer turns the test red. Dev-dependencies gain the three crates the
tests need through the public API; all were already in the lock.
